requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==4.2.9
djangorestframework==3.14.0
drf-spectacular==0.27.1
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
pkgutil_resolve_name==1.3.10
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
zipp==3.17.0
